Remove consented court helper and combine court list retrieval into one helper
